# public/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/todos":{"post":{"operationId":"Create Todo","summary":"","description":"Add new todo item","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":"Details of the newly created todo","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponseDto"}}}}},"tags":["Todos"]},"get":{"operationId":"All Todos","summary":"","description":"Fetch all todo items","parameters":[],"responses":{"200":{"description":"Details of all created todo items","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoResponseDto"}}}}}},"tags":["Todos"]}},"/api/todos/{id}":{"get":{"operationId":"Find By Id","summary":"","description":"Fetch information of a todo item","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Details of the todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponseDto"}}}}},"tags":["Todos"]},"put":{"operationId":"Update Todo","summary":"","description":"Update a todo item","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"200":{"description":"Updated details of the todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponseDto"}}}}},"tags":["Todos"]},"delete":{"operationId":"Delete Todo","summary":"","description":"Delete a todo item","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted details of the todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponseDto"}}}}},"tags":["Todos"]}}},"info":{"title":"Demo Credit API Docs","description":"API documentation for Demo Credit application","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateTodoDto":{"type":"object","properties":{"title":{"type":"string","description":"The title of the todo item","example":"Groceries shopping"},"body":{"type":"string","description":"The body of the todo item","example":"Buy groceries at 5pm today"}},"required":["title","body"]},"TodoResponseDto":{"type":"object","properties":{"title":{"type":"string","description":"The title of the todo item","example":"Groceries shopping"},"body":{"type":"string","description":"The body of the todo item","example":"Buy groceries at 5pm today"},"id":{"type":"string","description":"Unique id of the record","example":"638f9115ef73df6d34e3a3b2"},"updatedAt":{"type":"string","description":"Last date the record was updated in ISO3610 format","example":"\"2022-12-06T18:59:33.272Z"},"createdAt":{"type":"string","description":"Date the record was created in ISO3610 format","example":"\"2022-12-06T18:59:33.272Z"}},"required":["title","body","id","updatedAt","createdAt"]},"UpdateTodoDto":{"type":"object","properties":{"title":{"type":"string","description":"The title of the todo item","example":"Groceries shopping"},"body":{"type":"string","description":"The body of the todo item","example":"Buy groceries at 5pm today"}},"required":["title","body"]}}}}